move serde with_* helpers to their own module
