+++ crates/rethnet_eth/src/remote.rs
@@ -9,13 +9,16 @@ pub mod jsonrpc;
 /// RPC methods
 pub mod methods;
 
-mod withdrawal;
+/// helper utilities for use with serde's serialize_with and deserialize_with
+pub mod serde_with_helpers;
 
-use std::fmt::Write;
+mod withdrawal;
 
 use bytes::Bytes;
 
 use crate::U256;
+
+use serde_with_helpers::serialize_u256;
 
 pub use client::{RpcClient, RpcClientError};
 
@@ -34,74 +37,6 @@ impl From<u64> for U64 {
     fn from(u: u64) -> U64 {
         U64(u)
     }
-}
-
-/// for use with serde's serialize_with on a single value that should be serialized as a
-/// sequence
-pub fn single_to_sequence<S, T>(val: &T, s: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-    T: serde::Serialize,
-{
-    use serde::ser::SerializeSeq;
-    let mut seq = s.serialize_seq(Some(1))?;
-    seq.serialize_element(val)?;
-    seq.end()
-}
-
-/// for use with serde's deserialize_with on a sequence that should be deserialized as a
-/// single value
-pub fn sequence_to_single<'de, D, T>(deserializer: D) -> Result<T, D::Error>
-where
-    D: serde::Deserializer<'de>,
-    T: serde::Deserialize<'de> + Clone,
-{
-    let s: Vec<T> = serde::de::Deserialize::deserialize(deserializer)?;
-    Ok(s[0].clone())
-}
-
-/// a custom implementation because the one from ruint includes leading zeroes and the JSON-RPC
-/// server implementations reject that.
-fn serialize_u256<S>(x: &U256, s: S) -> Result<S::Ok, S::Error>
-where
-    S: serde::Serializer,
-{
-    let bytes = x.to_be_bytes_vec();
-
-    // OPT: Allocation free method.
-    let mut result = String::with_capacity(2 * U256::BYTES + 2);
-    result.push_str("0x");
-
-    let mut leading_zeroes = true;
-    for byte in bytes {
-        if leading_zeroes {
-            if byte != 0 {
-                write!(result, "{byte:x}").unwrap();
-                leading_zeroes = false;
-            }
-            continue;
-        }
-        write!(result, "{byte:02x}").unwrap();
-    }
-
-    // 0x0
-    if leading_zeroes {
-        result.push('0');
-    }
-
-    s.serialize_str(&result)
-}
-
-/// for use with serde's deserialize_with on fields of hexadecimal strings that should be
-/// parsed as Option<u64>
-pub fn optional_u64_from_hex<'de, D>(deserializer: D) -> Result<Option<u64>, D::Error>
-where
-    D: serde::Deserializer<'de>,
-{
-    let s: &str = serde::Deserialize::deserialize(deserializer)?;
-    Ok(Some(
-        u64::from_str_radix(&s[2..], 16).expect("failed to parse u64"),
-    ))
 }
 
 /// For specifying a block
